move bitvector-related expression classes into separate header file
std_expr.h is by far the largest header file, which impedes on readability.
This commit moves the bitvector-related expression classes into a separate
header file, following the pattern in mathematical_expr.h.
