+++ jbmc/src/java_bytecode/character_refine_preprocess.cpp
@@ -16,6 +16,7 @@ Date:   March 2017
 #include "character_refine_preprocess.h"
 
 #include <util/arith_tools.h>
+#include <util/bitvector_expr.h>
 #include <util/std_expr.h>
 
 /// converts based on a function on expressions
