+++ src/ansi-c/c_typecheck_base.h
@@ -12,12 +12,13 @@ Author: Daniel Kroening, kroening@kroening.com
 #ifndef CPROVER_ANSI_C_C_TYPECHECK_BASE_H
 #define CPROVER_ANSI_C_C_TYPECHECK_BASE_H
 
-#include <util/symbol_table.h>
-#include <util/typecheck.h>
+#include <util/bitvector_expr.h>
 #include <util/namespace.h>
 #include <util/std_code.h>
 #include <util/std_expr.h>
 #include <util/std_types.h>
+#include <util/symbol_table.h>
+#include <util/typecheck.h>
 
 #include "ansi_c_declaration.h"
 #include "designator.h"
