+++ src/solvers/flattening/boolbv_extractbit.cpp
@@ -11,6 +11,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <algorithm>
 
 #include <util/arith_tools.h>
+#include <util/bitvector_expr.h>
 #include <util/exception_utils.h>
 #include <util/std_expr.h>
 #include <util/std_types.h>
