+++ src/solvers/flattening/boolbv_extractbits.cpp
@@ -9,6 +9,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "boolbv.h"
 
 #include <util/arith_tools.h>
+#include <util/bitvector_expr.h>
 
 bvt boolbvt::convert_extractbits(const extractbits_exprt &expr)
 {
