+++ src/solvers/flattening/boolbv_replication.cpp
@@ -9,6 +9,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "boolbv.h"
 
 #include <util/arith_tools.h>
+#include <util/bitvector_expr.h>
 
 bvt boolbvt::convert_replication(const replication_exprt &expr)
 {
