+++ src/solvers/floatbv/float_bv.cpp
@@ -10,8 +10,9 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <algorithm>
 
-#include <util/std_expr.h>
 #include <util/arith_tools.h>
+#include <util/bitvector_expr.h>
+#include <util/std_expr.h>
 
 exprt float_bvt::convert(const exprt &expr) const
 {
