+++ src/solvers/qbf/qdimacs_core.cpp
@@ -9,6 +9,7 @@ Author: CM Wintersteiger
 #include "qdimacs_core.h"
 
 #include <util/arith_tools.h>
+#include <util/bitvector_expr.h>
 #include <util/std_expr.h>
 
 void qdimacs_coret::simplify_extractbits(exprt &expr) const
