+++ src/util/Makefile
@@ -3,6 +3,7 @@ SRC = allocate_objects.cpp \
       array_element_from_pointer.cpp \
       array_name.cpp \
       base_type.cpp \
+      bitvector_expr.cpp \
       bv_arithmetic.cpp \
       byte_operators.cpp \
       c_types.cpp \
