+++ src/util/simplify_expr.cpp
@@ -10,6 +10,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <algorithm>
 
+#include "bitvector_expr.h"
 #include "byte_operators.h"
 #include "c_types.h"
 #include "config.h"
