+++ unit/util/simplify_expr.cpp
@@ -9,6 +9,7 @@ Author: Michael Tautschnig
 #include <testing-utils/use_catch.h>
 
 #include <util/arith_tools.h>
+#include <util/bitvector_expr.h>
 #include <util/byte_operators.h>
 #include <util/c_types.h>
 #include <util/cmdline.h>
